src/requirements.txt
# Adafruit Libraries
Adafruit-Blinka==8.47.0
Adafruit-BME280==1.0.1
adafruit-circuitpython-bme280==2.6.25
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.11.0
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11

# GPIO and Hardware Control
binho-host-adapter==0.1.6
gpiozero==2.0.1
lgpio==0.2.2.0
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
spidev==3.6
sysv-ipc==1.1.0

# Data Processing and Analysis
numpy==1.26.4
pandas==2.2.3
pytz==2024.2
tzdata==2024.2

# Visualization
matplotlib==3.8.4
pillow==10.4.0
shapely==2.0.6

# Geospatial Libraries
basemap==1.4.1
basemap-data==1.3.2
Cartopy==0.23.0
geopandas==1.0.1
pyogrio==0.9.0
pyproj==3.6.1
pyshp==2.3.1

# Web and Networking
requests==2.32.3
httpx==0.27.2
httpcore==1.0.5
h11==0.14.0
sniffio==1.3.1
anyio==4.6.0

# Django and Web Development
Django==5.1.1
asgiref==3.8.1
sqlparse==0.5.1

# Fonts and Graphics
fonttools==4.54.1
pyfont==0.0.3.20041618
pyfonts==0.0.2
cairocffi==1.7.1
CairoSVG==2.7.1
webencodings==0.5.1

# Miscellaneous
certifi==2024.8.30
charset-normalizer==3.3.2
colorzero==2.0
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
idna==3.10
kiwisolver==1.4.7
packaging==23.2
pycparser==2.22
pyparsing==3.1.4
six==1.16.0
tinycss2==1.3.0
urllib3==2.2.3
